+++ app/routes/components/Tooltip/changelog.codegen.ts
@@ -7,4 +7,6 @@ export const changelog = ''
   + '- 2025-10-10: Move styles into `components` CSS layer. Code formatting\n'
   + '  improvement.\n'
   + '- 2025-10-31: Code formatting improvement.\n'
+  + '- 2025-11-27: Support the `anchor` prop to make the tooltip follow the mouse\n'
+  + '  cursor.\n'
   ;