+++ src/components/Tooltip/tooltip.tsx
@@ -1,7 +1,8 @@
-import type { FC, ReactNode } from "react";
+import classNames from "classnames";
+import { type FC, type ReactNode, useMemo } from "react";
 
+import { useJss, useTheme } from "@/components/CopyUiProvider";
 import { type Placement, Popover } from "@/components/Popover";
-import { useTheme } from "@/components/ThemeProvider";
 
 import styles from "./tooltip.module.scss";
 
@@ -14,22 +15,31 @@ interface TooltipTriggerRenderProps {
 
 interface TooltipProps {
   label: string;
+  anchor?: "element" | "pointer";
   placement?: Placement;
   triggerRender: (props: TooltipTriggerRenderProps) => ReactNode;
 }
 
 export const Tooltip: FC<TooltipProps> = (props) => {
-  const { label, placement = "top", triggerRender } = props;
+  const { label, anchor = "element", placement = "top", triggerRender } = props;
 
   const theme = useTheme();
+  const jss = useJss();
 
-  const computedStyle = {
+  const computedStyle = jss.hash({
     "--tooltip-background-color": theme.colors.gray["900"],
     "--tooltip-color": theme.colors.gray["100"],
-  };
+  });
+
+  const offset = useMemo(() => {
+    if (anchor === "pointer") {
+      return 12;
+    }
+    return undefined;
+  }, [anchor]);
 
   return (
-    <Popover placement={placement}>
+    <Popover offset={offset} anchor={anchor} placement={placement}>
       <Popover.Trigger render={triggerRender} />
       <Popover.Portal
         onClickOutside={({ closePortal }) => closePortal()}
@@ -37,8 +47,8 @@ export const Tooltip: FC<TooltipProps> = (props) => {
           isOpen && (
             <div
               ref={setRef}
-              className={styles.tooltip}
-              style={{ ...computedStyle, ...floatingStyles }}
+              className={classNames(styles.tooltip, computedStyle)}
+              style={floatingStyles}
             >
               {label}
             </div>