+++ src/components/Tooltip/CHANGELOG.md
@@ -3,3 +3,5 @@
 - 2025-10-10: Move styles into `components` CSS layer. Code formatting
   improvement.
 - 2025-10-31: Code formatting improvement.
+- 2025-11-27: Support the `anchor` prop to make the tooltip follow the mouse
+  cursor.